feat: Withdraw implementation for transactions

+++ tiny-bank-api/src/test/kotlin/com/tnfigueiredo/tinybank/bdd/steps/TransactionStepsDefinition.kt
@@ -3,6 +3,8 @@ package com.tnfigueiredo.tinybank.bdd.steps
 import com.tnfigueiredo.tinybank.model.*
 import com.tnfigueiredo.tinybank.model.DocType.NATIONAL_ID
 import com.tnfigueiredo.tinybank.model.TransactionType.DEPOSIT
+import com.tnfigueiredo.tinybank.model.TransactionType.WITHDRAW
+import io.cucumber.java.en.And
 import io.cucumber.java.en.Given
 import io.cucumber.java.en.Then
 import io.cucumber.java.en.When
@@ -25,7 +27,7 @@ class TransactionStepsDefinition {
         const val USERS_BASE_SERVICE_PATH = "/users"
         const val ACCOUNT_BASE_SERVICE_PATH = "/accounts"
         var accountToBeUsed: Account? = null
-        lateinit var deactivatedAccountToBeUsed: Account
+        var deactivatedAccountToBeUsed: Account? = null
         lateinit var result: ResponseEntity<RestResponse>
         var A_RANDOM_USER_ID: UUID = UUID.fromString("eae467d9-deb2-49b3-aaf5-f1e146e567e1")
         var ANOTHER_RANDOM_USER_ID: UUID = UUID.fromString("0865ef31-9e22-45e7-a8fd-0d4a815f8fe7")
@@ -66,45 +68,63 @@ class TransactionStepsDefinition {
 
     @Given("the user has a deactivated account in Tyny Bank")
     fun the_user_has_a_deactivated_account_in_tyny_bank() {
-        val anotherUserResult = restTemplate.postForEntity(USERS_BASE_SERVICE_PATH,
-            UserDTO(name = "ANOTHER_NAME", surname = "ANOTHER_SURNAME", docType = NATIONAL_ID, document = "ANOTHER_DOCUMENT", docCountry = "ANOTHER_DOC_COUNTRY"),
-            RestResponse::class.java
-        )
 
-        anotherUserResult.statusCode shouldBeEqual HttpStatus.OK
+        if (deactivatedAccountToBeUsed == null) {
+            val anotherUserResult = restTemplate.postForEntity(USERS_BASE_SERVICE_PATH,
+                UserDTO(name = "ANOTHER_NAME", surname = "ANOTHER_SURNAME", docType = NATIONAL_ID, document = "ANOTHER_DOCUMENT", docCountry = "ANOTHER_DOC_COUNTRY"),
+                RestResponse::class.java
+            )
 
-        ANOTHER_RANDOM_USER_ID = UUID.fromString(((anotherUserResult.body!!.data as Map<*, *>)["id"] as String))
-        val userDocType = DocType.valueOf(((anotherUserResult.body!!.data as Map<*, *>)["docType"] as String))
-        val document = ((anotherUserResult.body!!.data as Map<*, *>)["document"] as String)
-        val docCountry = ((anotherUserResult.body!!.data as Map<*, *>)["docCountry"] as String)
+            anotherUserResult.statusCode shouldBeEqual HttpStatus.OK
 
-        restTemplate.postForEntity(
-            "${ACCOUNT_BASE_SERVICE_PATH}/user/${ANOTHER_RANDOM_USER_ID}/agency/$AGENCY",
-            null,
-            RestResponse::class.java
-        ).statusCode shouldBe HttpStatus.OK
+            ANOTHER_RANDOM_USER_ID = UUID.fromString(((anotherUserResult.body!!.data as Map<*, *>)["id"] as String))
+            val userDocType = DocType.valueOf(((anotherUserResult.body!!.data as Map<*, *>)["docType"] as String))
+            val document = ((anotherUserResult.body!!.data as Map<*, *>)["document"] as String)
+            val docCountry = ((anotherUserResult.body!!.data as Map<*, *>)["docCountry"] as String)
 
-        restTemplate.delete(
-            "${USERS_BASE_SERVICE_PATH}/docType/${userDocType}/document/${document}/country/${docCountry}",
-            RestResponse::class.java
-        )
+            restTemplate.postForEntity(
+                "${ACCOUNT_BASE_SERVICE_PATH}/user/${ANOTHER_RANDOM_USER_ID}/agency/$AGENCY",
+                null,
+                RestResponse::class.java
+            ).statusCode shouldBe HttpStatus.OK
 
-        val deactivatedUserAndAccountResult = restTemplate.getForEntity(
-            "${USERS_BASE_SERVICE_PATH}/docType/${userDocType}/document/${document}/country/${docCountry}",
-            RestResponse::class.java
-        )
+            restTemplate.delete(
+                "${USERS_BASE_SERVICE_PATH}/docType/${userDocType}/document/${document}/country/${docCountry}",
+                RestResponse::class.java
+            )
 
-        val account = ((deactivatedUserAndAccountResult.body!!.data as Map<*, *>)["account"] as Map<*, *>)
-        deactivatedAccountToBeUsed = Account(
-            id = account["id"] as String,
-            agency = account["agency"].toString().toShort(),
-            year = account["year"].toString().toShort(),
-            userId = UUID.fromString(account["userId"] as String),
-            balance = account["balance"] as Double,
-            status = ActivationStatus.valueOf(account["status"] as String)
-        )
+            val deactivatedUserAndAccountResult = restTemplate.getForEntity(
+                "${USERS_BASE_SERVICE_PATH}/docType/${userDocType}/document/${document}/country/${docCountry}",
+                RestResponse::class.java
+            )
+
+            val account = ((deactivatedUserAndAccountResult.body!!.data as Map<*, *>)["account"] as Map<*, *>)
+            deactivatedAccountToBeUsed = Account(
+                id = account["id"] as String,
+                agency = account["agency"].toString().toShort(),
+                year = account["year"].toString().toShort(),
+                userId = UUID.fromString(account["userId"] as String),
+                balance = account["balance"] as Double,
+                status = ActivationStatus.valueOf(account["status"] as String)
+            )
+        }
 
         flowActiveUSer = false
+    }
+
+    @And("the account balance is {string}")
+    fun the_account_balance_is(depositValue: String) {
+        val accountUpdate = restTemplate.postForEntity(
+            BASE_SERVICE_PATH,
+            Transaction(
+                originAccountId = accountToBeUsed?.id!!,
+                amount = depositValue.toDouble(),
+                type = DEPOSIT
+            ),
+            RestResponse::class.java
+        )
+
+        accountToBeUsed = accountToBeUsed!!.copy(balance = accountToBeUsed!!.balance + depositValue.toDouble())
     }
 
     @When("the user makes a deposit of {string}")
@@ -124,9 +144,36 @@ class TransactionStepsDefinition {
             restTemplate.postForEntity(
                 BASE_SERVICE_PATH,
                 Transaction(
-                    originAccountId = deactivatedAccountToBeUsed.id!!,
+                    originAccountId = deactivatedAccountToBeUsed?.id!!,
                     amount = transactionValue.toDouble(),
                     type = DEPOSIT
+                ),
+                RestResponse::class.java
+            )
+        }
+
+    }
+
+    @When("the user makes a withdraw of {string}")
+    fun the_user_makes_a_withdraw_of(transactionValue: String) {
+        result = if(flowActiveUSer) {
+            restTemplate.postForEntity(
+                BASE_SERVICE_PATH,
+                Transaction(
+                    originAccountId = accountToBeUsed?.id!!,
+                    amount = transactionValue.toDouble(),
+                    type = WITHDRAW
+                ),
+                RestResponse::class.java
+            )
+
+        } else {
+            restTemplate.postForEntity(
+                BASE_SERVICE_PATH,
+                Transaction(
+                    originAccountId = deactivatedAccountToBeUsed?.id!!,
+                    amount = transactionValue.toDouble(),
+                    type = WITHDRAW
                 ),
                 RestResponse::class.java
             )
@@ -139,6 +186,15 @@ class TransactionStepsDefinition {
         result.statusCode shouldBeEqual HttpStatus.OK
         (result.body!!.data as Map<*, *>)["id"].shouldNotBeNull()
         Serenity.recordReportData().withTitle("Client Account Initial State").andContents(accountToBeUsed.toString())
+        Serenity.recordReportData().withTitle("Transaction Final State").andContents(result.toString())
+    }
+
+    @Then("the account balance is decreased by {string}")
+    fun the_account_balance_is_decreased_by(string: String?) {
+        result.statusCode shouldBeEqual HttpStatus.OK
+        (result.body!!.data as Map<*, *>)["id"].shouldNotBeNull()
+        Serenity.recordReportData().withTitle("Client Account Initial State").andContents(accountToBeUsed.toString())
+        Serenity.recordReportData().withTitle("Transaction Final State").andContents(result.toString())
     }
 
     @Then("the transaction operation fails")
